refactor: response 생성 책임 분리
