+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11;
 
 import com.techcourse.exception.UncheckedServletException;
-import com.techcourse.model.HttpRequest;
-import com.techcourse.model.RequestToResponse;
+import org.apache.coyote.http.HttpRequest;
+import org.apache.coyote.http.RequestToResponse;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -53,8 +53,6 @@ public class Http11Processor implements Runnable, Processor {
             request.add(bufferedReader.readLine());
         }
 
-        HttpRequest httpRequest = HttpRequest.of(request);
-
-        return httpRequest;
+        return HttpRequest.of(request);
     }
 }
