+++ tomcat/src/test/java/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -51,8 +51,9 @@ class Http11ProcessorTest {
         // then
         final URL resource = getClass().getClassLoader().getResource("static/index.html");
         var expected = "HTTP/1.1 200 OK \r\n" +
-                "Content-Type: text/html;charset=utf-8 \r\n" +
+
                 "Content-Length: 5564 \r\n" +
+                "Content-Type: text/html;charset=utf-8 \r\n" +
                 "\r\n"+
                 new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
